+++ packages/plugins/src/plugins/diagnostic-comment.plugin.ts
@@ -10,6 +10,13 @@ import {
 import { findNodeAtPosition } from '@rehearsal/ts-utils';
 import debug from 'debug';
 import ts, { LanguageService } from 'typescript';
+import {
+  findDiagnosticNode,
+  getConditionalCommentPos,
+  inJsxText,
+  inTemplateExpressionText,
+  onMultilineConditionalTokenLine,
+} from './utils.js';
 import type MS from 'magic-string';
 
 const require = createRequire(import.meta.url);
@@ -44,7 +51,7 @@ export class DiagnosticCommentPlugin implements Plugin<DiagnosticCommentPluginOp
 
     const diagnostics = this.getDiagnostics(context.service, fileName, options.commentTag);
 
-    DEBUG_CALLBACK(`Plugin 'DiagnosticReport' run on %O:`, fileName);
+    DEBUG_CALLBACK(`Plugin 'DiagnosticCommentPlugin' run on %O:`, fileName);
 
     const allFixedFiles: Set<string> = new Set();
 
@@ -179,7 +186,7 @@ export class DiagnosticCommentPlugin implements Plugin<DiagnosticCommentPluginOp
     const lineAndChar = getLineAndCharacterOfPosition(diagnostic.file, diagnostic.start);
 
     /**
-     * The following logic has been ported from  https://github.com/airbnb/ts-migrate/blob/master/packages/ts-migrate-plugins/src/plugins/ts-ignore.ts#L199
+     * The following logic has been ported from https://github.com/airbnb/ts-migrate/blob/master/packages/ts-migrate-plugins/src/plugins/ts-ignore.ts#L199 and modified for our use cases
      *
      * It correctly finds the valid place to stick the inline comment
      */
@@ -228,138 +235,4 @@ export class DiagnosticCommentPlugin implements Plugin<DiagnosticCommentPluginOp
   isErrorDiagnostic(diagnostic: DiagnosticWithContext): boolean {
     return diagnostic.category === DiagnosticCategory.Error;
   }
-}
-
-function inTemplateExpressionText(sourceFile: ts.SourceFile, pos: number): boolean {
-  const visitor = (node: ts.Node): boolean | undefined => {
-    if (node.pos <= pos && pos < node.end && ts.isTemplateExpression(node)) {
-      const inHead = node.head.pos <= pos && pos < node.head.end;
-      const inMiddleOrTail = node.templateSpans.some(
-        (span) => span.literal.pos <= pos && pos < span.literal.end
-      );
-      if (inHead || inMiddleOrTail) {
-        return true;
-      }
-    }
-
-    return ts.forEachChild(node, visitor);
-  };
-
-  return !!ts.forEachChild(sourceFile, visitor);
-}
-
-function findDiagnosticNode(
-  diagnostic: ts.DiagnosticWithLocation,
-  sourceFile: ts.SourceFile
-): ts.Node | undefined {
-  const visitor = (node: ts.Node): ts.Node | undefined =>
-    isDiagnosticNode(node, diagnostic, sourceFile) ? node : ts.forEachChild(node, visitor);
-
-  return visitor(sourceFile);
-}
-
-function isDiagnosticNode(
-  node: ts.Node,
-  diagnostic: ts.DiagnosticWithLocation,
-  sourceFile: ts.SourceFile
-): boolean {
-  return (
-    node.getStart(sourceFile) === diagnostic.start &&
-    node.getEnd() === diagnostic.start + diagnostic.length
-  );
-}
-
-function inJsxText(sourceFile: ts.SourceFile, pos: number): boolean {
-  const visitor = (node: ts.Node): boolean | undefined => {
-    if (node.pos <= pos && pos < node.end && (ts.isJsxElement(node) || ts.isJsxFragment(node))) {
-      const isJsxTextChild = node.children.some(
-        (child) => ts.isJsxText(child) && child.pos <= pos && pos < child.end
-      );
-      const isClosingElement = !ts.isJsxFragment(node) && node.closingElement.pos === pos;
-      if (isJsxTextChild || isClosingElement) {
-        return true;
-      }
-    }
-
-    return ts.forEachChild(node, visitor);
-  };
-
-  return !!ts.forEachChild(sourceFile, visitor);
-}
-
-function onMultilineConditionalTokenLine(sourceFile: ts.SourceFile, pos: number): boolean {
-  const conditionalExpression = getConditionalExpressionAtPos(sourceFile, pos);
-  // Not in a conditional expression.
-  if (!conditionalExpression) {
-    return false;
-  }
-
-  const { line: questionTokenLine } = ts.getLineAndCharacterOfPosition(
-    sourceFile,
-    conditionalExpression.questionToken.end
-  );
-  const { line: colonTokenLine } = ts.getLineAndCharacterOfPosition(
-    sourceFile,
-    conditionalExpression.colonToken.end
-  );
-  // Single line conditional expression.
-  if (questionTokenLine === colonTokenLine) {
-    return false;
-  }
-
-  const { line } = ts.getLineAndCharacterOfPosition(sourceFile, pos);
-  return visitConditionalExpressionWhen(conditionalExpression, pos, {
-    // On question token line of multiline conditional expression.
-    whenTrue: () => line === questionTokenLine,
-    // On colon token line of multiline conditional expression.
-    whenFalse: () => line === colonTokenLine,
-    otherwise: () => false,
-  });
-}
-
-function getConditionalExpressionAtPos(
-  sourceFile: ts.SourceFile,
-  pos: number
-): ts.ConditionalExpression | undefined {
-  const visitor = (node: ts.Node): ts.ConditionalExpression | undefined => {
-    if (node.pos <= pos && pos < node.end && ts.isConditionalExpression(node)) {
-      return node;
-    }
-    return ts.forEachChild(node, visitor);
-  };
-  return ts.forEachChild(sourceFile, visitor);
-}
-
-function visitConditionalExpressionWhen<T>(
-  node: ts.ConditionalExpression | undefined,
-  pos: number,
-  visitor: {
-    whenTrue(node: ts.ConditionalExpression): T;
-    whenFalse(node: ts.ConditionalExpression): T;
-    otherwise(): T;
-  }
-): T {
-  if (!node) {
-    return visitor.otherwise();
-  }
-
-  const inWhenTrue = node.whenTrue.pos <= pos && pos < node.whenTrue.end;
-  if (inWhenTrue) {
-    return visitor.whenTrue(node);
-  }
-
-  const inWhenFalse = node.whenFalse.pos <= pos && pos < node.whenFalse.end;
-  if (inWhenFalse) {
-    return visitor.whenFalse(node);
-  }
-
-  return visitor.otherwise();
-}
-
-function getConditionalCommentPos(sourceFile: ts.SourceFile, pos: number): number {
-  return visitConditionalExpressionWhen(getConditionalExpressionAtPos(sourceFile, pos), pos, {
-    whenTrue: (node) => node.questionToken.end,
-    whenFalse: (node) => node.colonToken.end,
-    otherwise: () => pos,
-  });
 }
